+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/SpringCoreClasses.kt
@@ -9,6 +9,11 @@ object SpringCoreClasses {
     const val SPRING_BOOT_APPLICATION = "org.springframework.boot.autoconfigure.SpringBootApplication"
     const val ACTUATOR_ENDPOINT = "org.springframework.boot.actuate.endpoint.annotation.Endpoint"
     const val ACTUATOR_ENDPOINT_ACCESS = "org.springframework.boot.actuate.endpoint.Access"
+    const val ACTUATOR_JMX_ENDPOINT = "org.springframework.boot.actuate.endpoint.jmx.annotation.JmxEndpoint"
+    const val ACTUATOR_READ_OPERATION = "org.springframework.boot.actuate.endpoint.annotation.ReadOperation"
+    const val ACTUATOR_WRITE_OPERATION = "org.springframework.boot.actuate.endpoint.annotation.WriteOperation"
+    const val ACTUATOR_DELETE_OPERATION = "org.springframework.boot.actuate.endpoint.annotation.DeleteOperation"
+    const val ACTUATOR_SELECTOR = "org.springframework.boot.actuate.endpoint.annotation.Selector"
     const val SPRING_APPLICATION = "org.springframework.boot.SpringApplication"
     const val BOOT_AUTO_CONFIGURATION = "org.springframework.boot.autoconfigure.AutoConfiguration"